% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{b}{}
}
\value{
graph
}
\description{
Central Limit function
}
\examples{
mycltu(n=20,iter=100000)
}
